requirements.txt
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
fake-useragent==1.1.3
filelock==3.12.2
greenlet==2.0.1
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
install==1.3.5
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
packaging==23.1
parsel==1.8.1
playwright==1.29.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
tldextract==3.4.4
tomli==2.0.1
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==2.0.3
w3lib==2.1.1
zipp==3.15.0
zope.interface==6.0
